# front/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "新闻资讯发布与推荐系统 API文档",
    "description": "我的新闻资讯系统后台接口文档",
    "contact": {
      "name": "Tong",
      "email": "tong@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/system/category": {
      "put": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "edit_11",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "add_11",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/export": {
      "post": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "export_7",
        "parameters": [
          {
            "name": "news",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/News"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/system/category/export": {
      "post": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "export_11",
        "parameters": [
          {
            "name": "newsCategory",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NewsCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/system/user/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getInfo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/tagRel/{id}": {
      "get": {
        "tags": [
          "news-tag-rel-controller"
        ],
        "operationId": "getInfo_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/tag/{id}": {
      "get": {
        "tags": [
          "news-tag-controller"
        ],
        "operationId": "getInfo_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/sync/{newsId}": {
      "get": {
        "tags": [
          "redis-sync-temp-controller"
        ],
        "operationId": "getInfo_3",
        "parameters": [
          {
            "name": "newsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/statistics/{id}": {
      "get": {
        "tags": [
          "news-statistics-controller"
        ],
        "operationId": "getInfo_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/role/{id}": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "getInfo_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/operlog/{id}": {
      "get": {
        "tags": [
          "sys-operation-log-controller"
        ],
        "operationId": "getInfo_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/{id}": {
      "get": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "getInfo_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/list": {
      "get": {
        "tags": [
          "新闻资讯"
        ],
        "summary": "查询新闻列表(管理端)",
        "operationId": "list_7",
        "parameters": [
          {
            "name": "news",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/News"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataInfo"
                }
              }
            }
          }
        }
      }
    },
    "/system/like/{id}": {
      "get": {
        "tags": [
          "news-like-controller"
        ],
        "operationId": "getInfo_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/config/{id}": {
      "get": {
        "tags": [
          "sys-config-controller"
        ],
        "operationId": "getInfo_9",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/comment/{id}": {
      "get": {
        "tags": [
          "评论管理"
        ],
        "operationId": "getInfo_10",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/category/{id}": {
      "get": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "getInfo_11",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/category/list": {
      "get": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "list_11",
        "parameters": [
          {
            "name": "newsCategory",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NewsCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/browse/{id}": {
      "get": {
        "tags": [
          "user-browse-controller"
        ],
        "operationId": "getInfo_12",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/news/{ids}": {
      "delete": {
        "tags": [
          "新闻资讯"
        ],
        "operationId": "remove_7",
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/category/{ids}": {
      "delete": {
        "tags": [
          "news-category-controller"
        ],
        "operationId": "remove_11",
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AjaxResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          },
          "warn": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        },
        "additionalProperties": {
          "type": "object"
        }
      },
...
      "News": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateBy": {
            "type": "string",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "remark": {
            "type": "string",
            "refType": null
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "refType": null
          },
          "id": {
            "type": "integer",
            "description": "主键ID",
            "format": "int64",
            "refType": null
          },
          "title": {
            "type": "string",
            "description": "新闻标题",
            "refType": null
          },
          "summary": {
            "type": "string",
            "description": "新闻摘要",
            "refType": null
          },
          "categoryId": {
            "type": "integer",
            "description": "所属分类ID",
            "format": "int64",
            "refType": null
          },
          "coverUrl": {
            "type": "string",
            "refType": null
          },
          "content": {
            "type": "string",
            "refType": null
          },
          "publishStatus": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "shelfStatus": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "publishUserId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "publishTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "isDeleted": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "categoryName": {
            "type": "string",
            "refType": null
          }
        },
        "description": "新闻资讯对象"
      },
...
      "NewsCategory": {
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateBy": {
            "type": "string",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "remark": {
            "type": "string",
            "refType": null
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "refType": null
          },
          "parentName": {
            "type": "string",
            "refType": null
          },
          "parentId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "orderNum": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "ancestors": {
            "type": "string",
            "refType": null
          },
          "children": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "refType": "object"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "categoryName": {
            "type": "string",
            "refType": null
          },
          "sort": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "isDeleted": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
...
      "TableDataInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          }
        }
      }
    }
  }
}